+++ awx/ui_next/src/screens/Project/ProjectAdd/ProjectAdd.test.jsx
@@ -37,6 +37,7 @@ describe('<ProjectAdd />', () => {
               ['git', 'Git'],
               ['hg', 'Mercurial'],
               ['svn', 'Subversion'],
+              ['archive', 'Remote Archive'],
               ['insights', 'Red Hat Insights'],
             ],
           },